Swap controller names: specimen module controller now maps to "specimen"
