+++ specimen/src/org/labkey/specimen/view/manageActorOrder.jsp
@@ -17,8 +17,8 @@
 %>
 <%@ page import="org.labkey.api.specimen.model.SpecimenRequestActor"%>
 <%@ page import="org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider"%>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageActorOrderAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageActorsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorOrderAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorsAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
