+++ specimen/src/org/labkey/specimen/view/manageActors.jsp
@@ -25,9 +25,9 @@
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.specimen.actions.ShowGroupMembersAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.DeleteActorAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageActorOrderAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageActorsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.DeleteActorAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorOrderAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorsAction" %>
 <%@ page import="java.util.Set" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
