+++ specimen/src/org/labkey/specimen/view/manageComments.jsp
@@ -26,8 +26,8 @@
 <%@ page import="org.labkey.api.study.TimepointType" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageCommentsForm" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageSpecimenCommentsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageCommentsForm" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageSpecimenCommentsAction" %>
 <%@ page import="java.util.Collections" %>
 <%@ page import="java.util.List" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
