+++ specimen/src/org/labkey/specimen/view/manageDisplay.jsp
@@ -19,7 +19,7 @@
 <%@ page import="org.labkey.api.study.StudyUrls"%>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageDisplaySettingsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageDisplaySettingsAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
