+++ specimen/src/org/labkey/specimen/view/manageNotifications.jsp
@@ -30,7 +30,7 @@
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageNotificationsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageNotificationsAction" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
