+++ specimen/src/org/labkey/specimen/view/manageRequestInputs.jsp
@@ -20,7 +20,7 @@
 <%@ page import="org.labkey.api.view.HttpView"%>
 <%@ page import="org.labkey.api.view.JspView"%>
 <%@ page import="org.labkey.specimen.actions.ManageRequestInputsBean"%>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.HandleUpdateRequestInputsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.HandleUpdateRequestInputsAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
