+++ specimen/src/org/labkey/specimen/view/manageSpecimens.jsp
@@ -30,17 +30,17 @@
 <%@ page import="org.labkey.api.study.Study" %>
 <%@ page import="org.labkey.api.study.StudyService" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ChooseImporterAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ConfigureRequestabilityRulesAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageActorsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageDefaultReqsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageDisplaySettingsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageLocationTypesAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageNotificationsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageRepositorySettingsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageRequestInputsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageSpecimenCommentsAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageStatusesAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ChooseImporterAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ConfigureRequestabilityRulesAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageActorsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageDefaultReqsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageDisplaySettingsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageLocationTypesAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageNotificationsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageRepositorySettingsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageRequestInputsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageSpecimenCommentsAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusesAction" %>
 <%@ page import="java.util.Collection" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
