+++ specimen/src/org/labkey/specimen/view/manageStatusOrder.jsp
@@ -21,9 +21,9 @@
 <%@ page import="org.labkey.api.view.HttpView"%>
 <%@ page import="org.labkey.api.view.JspView"%>
 <%@ page import="java.util.List" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageStatusOrderAction" %>
-<%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageStatusesAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusOrderAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusesAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
