--- study/src/org/labkey/study/controllers/specimen/SpecimenController.java
@@ -1,323 +1,0 @@
-/*
- * Copyright (c) 2014-2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.labkey.study.controllers.specimen;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.labkey.api.action.SimpleViewAction;
-import org.labkey.api.data.BaseColumnInfo;
-import org.labkey.api.data.ButtonBar;
-import org.labkey.api.data.Container;
-import org.labkey.api.data.SimpleFilter;
-import org.labkey.api.data.TableInfo;
-import org.labkey.api.query.FieldKey;
-import org.labkey.api.query.QueryParam;
-import org.labkey.api.query.QueryUpdateForm;
-import org.labkey.api.query.QueryView;
-import org.labkey.api.query.UserSchemaAction;
-import org.labkey.api.security.RequiresPermission;
-import org.labkey.api.security.permissions.ReadPermission;
-import org.labkey.api.specimen.SpecimenMigrationService;
-import org.labkey.api.specimen.actions.ParticipantCommentForm;
-import org.labkey.api.specimen.actions.SpecimenViewTypeForm;
-import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.study.SpecimenUrls;
-import org.labkey.api.study.StudyService;
-import org.labkey.api.study.StudyUrls;
-import org.labkey.api.view.ActionURL;
-import org.labkey.api.view.DataView;
-import org.labkey.api.view.HtmlView;
-import org.labkey.api.view.HttpView;
-import org.labkey.api.view.InsertView;
-import org.labkey.api.view.NavTree;
-import org.labkey.api.view.UpdateView;
-import org.labkey.study.controllers.BaseStudyController;
-import org.labkey.study.model.DatasetDefinition;
-import org.labkey.study.model.StudyImpl;
-import org.labkey.study.model.StudyManager;
-import org.labkey.study.query.DatasetQuerySettings;
-import org.labkey.study.query.DatasetQueryView;
-import org.labkey.study.query.SpecimenDetailTable;
-import org.labkey.study.query.StudyQuerySchema;
-import org.springframework.validation.BindException;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.Controller;
-
-import java.sql.ResultSet;
-
-/**
- * User: brittp
- * Date: Dec 20, 2007
- * Time: 11:08:31 AM
- */
-@SuppressWarnings("UnusedDeclaration")
-public class SpecimenController extends BaseStudyController
-{
-    private static final Logger _log = LogManager.getLogger(SpecimenController.class);
-
-    private static final DefaultActionResolver _actionResolver = new DefaultActionResolver(
-        SpecimenController.class,
-        ParticipantCommentAction.SpecimenCommentInsertAction.class,
-        ParticipantCommentAction.SpecimenCommentUpdateAction.class
-    );
-
-    public SpecimenController()
-    {
-        setActionResolver(_actionResolver);
-    }
-
-    @Override
-    protected void _addManageStudy(NavTree root)
-    {
-        urlProvider(StudyUrls.class).addManageStudyNavTrail(root, getContainer(), getUser());
-    }
-
-    public static class SpecimenUrlsImpl implements SpecimenUrls
-    {
-        @Override
-        public ActionURL getSpecimensURL(Container c)
-        {
-            return SpecimenMigrationService.get().getSpecimensURL(c);
-        }
-
-        @Override
-        public ActionURL getSpecimensURL(Container c, boolean showVials)
-        {
-            return getSpecimensURL(c).addParameter(SpecimenViewTypeForm.PARAMS.showVials, showVials);
-        }
-
-        @Override
-        public ActionURL getManageRequestURL(Container c, int requestId)
-        {
-            return SpecimenMigrationService.get().getManageRequestURL(c, requestId, null);
-        }
-
-        @Override
-        public ActionURL getRequestDetailsURL(Container c, int requestId)
-        {
-            return getManageRequestURL(c, requestId);
-        }
-
-        @Override
-        public ActionURL getManageRequestStatusURL(Container c, int requestId)
-        {
-            return SpecimenMigrationService.get().getManageRequestStatusURL(c, requestId);
-        }
-
-        @Override
-        public ActionURL getUpdateSpecimenQueryRowURL(Container c, String schemaName, TableInfo table)
-        {
-            ActionURL url = new ActionURL(UpdateSpecimenQueryRowAction.class, c);
-            url.addParameter("schemaName", schemaName);
-            url.addParameter(QueryView.DATAREGIONNAME_DEFAULT + "." + QueryParam.queryName, table.getName());
-
-            return url;
-        }
-
-        @Override
-        public ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table)
-        {
-            ActionURL url = new ActionURL(InsertSpecimenQueryRowAction.class, c);
-            url.addParameter("schemaName", schemaName);
-            url.addParameter(QueryView.DATAREGIONNAME_DEFAULT + "." + QueryParam.queryName, table.getName());
-
-            return url;
-        }
-
-        @Override
-        public Class<? extends Controller> getCopyParticipantCommentActionClass()
-        {
-            return CopyParticipantCommentAction.class;
-        }
-    }
-
-    @RequiresPermission(ReadPermission.class)
-    public class CopyParticipantCommentAction extends SimpleViewAction<ParticipantCommentForm>
-    {
-        @Override
-        public ModelAndView getView(final ParticipantCommentForm form, BindException errors)
-        {
-            StudyImpl study = getStudyRedirectIfNull();
-            DatasetDefinition def;
-
-            if (form.getVisitId() != 0)
-            {
-                def = StudyManager.getInstance().getDatasetDefinition(study, study.getParticipantVisitCommentDatasetId());
-            }
-            else
-            {
-                def = StudyManager.getInstance().getDatasetDefinition(study, study.getParticipantCommentDatasetId());
-            }
-
-            if (def != null)
-            {
-                StudyQuerySchema querySchema = StudyQuerySchema.createSchema(study, getUser(), true);
-                DatasetQuerySettings qs = (DatasetQuerySettings)querySchema.getSettings(getViewContext(), DatasetQueryView.DATAREGION, def.getName());
-                qs.setUseQCSet(false);
-
-                DatasetQueryView queryView = new DatasetQueryView(querySchema, qs, errors)
-                {
-                    @Override
-                    public DataView createDataView()
-                    {
-                        DataView view = super.createDataView();
-
-                        SimpleFilter filter = (SimpleFilter) view.getRenderContext().getBaseFilter();
-                        if (null == filter)
-                        {
-                            filter = new SimpleFilter();
-                            view.getRenderContext().setBaseFilter(filter);
-                        }
-                        filter.addCondition(StudyService.get().getSubjectColumnName(view.getViewContext().getContainer()), form.getParticipantId());
-                        if (form.getVisitId() != 0)
-                            filter.addCondition(FieldKey.fromParts("sequenceNum"), form.getVisitId());
-
-                        return view;
-                    }
-                };
-
-                String lsid = getExistingComment(queryView);
-                ActionURL url;
-                if (lsid != null)
-                    url = new ActionURL(ParticipantCommentAction.SpecimenCommentUpdateAction.class, getContainer()).
-                            addParameter(ParticipantCommentForm.params.lsid, lsid);
-                else
-                    url = new ActionURL(ParticipantCommentAction.SpecimenCommentInsertAction.class, getContainer()).
-                            addParameter(ParticipantCommentForm.params.participantId, form.getParticipantId());
-
-                url.addParameter(ParticipantCommentForm.params.datasetId, def.getDatasetId()).
-                        addParameter(ParticipantCommentForm.params.comment, form.getComment()).
-                        addReturnURL(form.getReturnActionURL()).
-                        addParameter(ParticipantCommentForm.params.visitId, form.getVisitId());
-
-                for (int rowId : form.getVialCommentsToClear())
-                    url.addParameter(ParticipantCommentForm.params.vialCommentsToClear, rowId);
-                return HttpView.redirect(url);
-            }
-            return new HtmlView("Dataset could not be found");
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-        }
-
-        private String getExistingComment(QueryView queryView)
-        {
-            final TableInfo table = queryView.getTable();
-            if (table != null)
-            {
-                try (ResultSet rs = queryView.getResultSet())
-                {
-                    while (rs.next())
-                    {
-                        String lsid = rs.getString("lsid");
-                        if (lsid != null)
-                            return lsid;
-                    }
-                }
-                catch (Exception e)
-                {
-                    _log.error("Error encountered trying to get " + StudyService.get().getSubjectNounSingular(getContainer()) + " comments", e);
-                }
-            }
-            return null;
-        }
-    }
-
-    @RequiresPermission(EditSpecimenDataPermission.class)
-    public static class UpdateSpecimenQueryRowAction extends UserSchemaAction
-    {
-        @Override
-        public ModelAndView getView(QueryUpdateForm tableForm, boolean reshow, BindException errors)
-        {
-            // Don't allow GlobalUniqueId to be edited
-            TableInfo tableInfo = tableForm.getTable();
-            if (null != tableInfo.getColumn("GlobalUniqueId"))
-                ((BaseColumnInfo)tableInfo.getColumn("GlobalUniqueId")).setReadOnly(true);
-
-            var vialComments = tableInfo.getColumn("VialComments");
-            if (null != vialComments)
-            {
-                ((BaseColumnInfo)vialComments).setUserEditable(false);
-                ((BaseColumnInfo)vialComments).setHidden(true);
-            }
-
-            SpecimenController.fixSpecimenRequestableColumn(tableForm);
-            ButtonBar bb = createSubmitCancelButtonBar(tableForm);
-            UpdateView view = new UpdateView(tableForm, errors);
-            view.getDataRegion().setButtonBar(bb);
-            return view;
-        }
-
-        @Override
-        public boolean handlePost(QueryUpdateForm tableForm, BindException errors)
-        {
-            doInsertUpdate(tableForm, errors, false);
-            return 0 == errors.getErrorCount();
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            super.addNavTrail(root);
-            root.addChild("Edit " + _form.getQueryName());
-        }
-    }
-
-    @RequiresPermission(EditSpecimenDataPermission.class)
-    public static class InsertSpecimenQueryRowAction extends UserSchemaAction
-    {
-        @Override
-        public ModelAndView getView(QueryUpdateForm tableForm, boolean reshow, BindException errors)
-        {
-            TableInfo tableInfo = tableForm.getTable();
-            var vialComments = tableInfo.getColumn("VialComments");
-            if (null != vialComments)
-            {
-                ((BaseColumnInfo)vialComments).setUserEditable(false);
-                ((BaseColumnInfo)vialComments).setHidden(true);
-            }
-
-            SpecimenController.fixSpecimenRequestableColumn(tableForm);
-            InsertView view = new InsertView(tableForm, errors);
-            view.getDataRegion().setButtonBar(createSubmitCancelButtonBar(tableForm));
-            return view;
-        }
-
-        @Override
-        public boolean handlePost(QueryUpdateForm tableForm, BindException errors)
-        {
-            doInsertUpdate(tableForm, errors, true);
-            return 0 == errors.getErrorCount();
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            super.addNavTrail(root);
-            root.addChild("Insert " + _form.getQueryName());
-        }
-    }
-
-    private static void fixSpecimenRequestableColumn(QueryUpdateForm tableForm)
-    {
-        TableInfo tableInfo = tableForm.getTable(); //TODO: finish fixing bug
-        if (tableInfo instanceof SpecimenDetailTable)
-            ((SpecimenDetailTable)tableInfo).changeRequestableColumn();
-    }
-}
